Add generate_seq2seq.py
